subsystem: lorawan: Put lorawan processing into its own thread
As the loramac-node library expects a function be run explicitly
to process all things non-time-critical it makes sense to put
it in its own thread freeing the library user from having to
call it in the app.

This commit also removes a blocking temporary code which used
to be used to debug join-accept messaging.

Signed-off-by: Kuba Sanak <contact@kuba.fyi>